[ci] remove dependency on docker CI base image (#1194)
Our dockerized CI tests currently run based off the `ci-base` image,
which is just a golang image with our directory in it. This is somewhat
inefficient as we have to bump another docker image if we want to change
the test harness.

This modifies the buildkite builds to use a similar pattern we've used
in other repos, which is to base directly off the golang image and add
the workdir via docker-compose configs.

Additionally, a script is included for creating buildkite builds in an
adhoc manner. The URL referenced in the script is publicly
reconstructable just by looking at any PR on the monorepo, so we're not
exposing any info.